engine: Support pagination for errata
The patch extends the api against Katello/Satellite and
supports filtering by erratum type and pagination.
The filtering criteria should be set via ErrataFilter
which stores the page number, page size and the requested
types of errata to query.

The result of a filtered request will contain ErrataData
object which consist the counts of the total count of errata
for an unfiltered query, the subtotal count of errata for the
filtered query and the errata bounded by the page size and
categorized by type and severity.

Change-Id: I2ff8a8b47ab210eec3e6d8fd6ed629655c650952
Signed-off-by: Moti Asayag <masayag@redhat.com>