core: Change min allocated memory of blank template
The min allocated memory of blank template was by default 0 which is not
correct and if such a template is edited, the validations fail.

Fixed by changing the min allocated memory to memory size for the blank
template if the min allocated memory was still 0 (do not touch if the user has
already changed it).

Change-Id: I860bf6567462b89260481d4435530295639a7c11
Bug-Url: https://bugzilla.redhat.com/1294488
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>