core: Fix of MigrateDowntimeSupported option
There was a typo in database upgrade script for the database option
MigrateDowntimeSupported in commit 1a2a5179.  As a result, the option
name is different in the source code and in the database, producing
error message in the log after otherwise successful live migration with
clusters < 3.6.

This commit fixes the typo in the database.

Change-Id: I6a1f88b85c561b6a6b2e5aa59d4e1aeea005ad6a
Bug-Url: https://bugzilla.redhat.com/1272084
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>