core: Change modifier of constructors to public
Since there was a refactor of the classes packages,
some commands which were used to be called as part of
the infrastructure flow fails to be called now,
since their constructors are protected and they were
moved to a different package then the CommandBase.
One example is removeImageCommand which use as part
of revert task of addVMTemaplate.

Change-Id: If1930ba0fccfa7f65234bf3d1e4bc953c73e5041
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>