backend: make commands' constructors accessible from CommandsFactory
Since CommandsFactory is the point where commands' constructors are
called, it should be able to access them.
Thus, each constructor's modifier of a command that is located:
- outside of CommandsFactory's package - should be public.
- in the same package as CommandsFactory - should not be private.

This patch adds a test to enforce this constraint, alongside changing
the modifiers of constructors that don't fulfill it.

Note that constructors which are used as a part of the compensation
process on startup (those that get the command's guid as a parameter)
are not exceptional, and should also fulfill this constraint.
Today, we call setAccessible(true) on these constructors before
accessing them. This concept is wrong for a few reasons:
1. It slows down the command's creation.
2. It's Java Security configuration dependant - a user/costumer can
   accidently break it.
3. It's not thread safe - today, the compensation process that runs on
   startup is serial. But if one day we would make it parallel,
   createCommand(className, commandId) will break.
   For example: thread A enters the "if" statement and sets the
   constructor's accessability to true. Then thread B doesn't enter the
   "if", thread A ends its excecution (after setting the constructor's
   accessability to false) and thread B throws NoSuchMethodException
   when trying to call getDeclaredConstructor.
4. There's no point in setting these constructors modifiers as
   protected, when we know they always should have public access.

Actually, there is no advantage in this approach, and thus this patch
removes the block of code that deals with the constructor's
accessability, counting on the test that it's already accessible.

Change-Id: Ic4598946f37f25b7883ee24719a99a75b3eb60b1
Signed-off-by: Idan Shaby <ishaby@redhat.com>