frontend: add missing warning label to FcpStorageView
This patch adds the warning label that already existed in the model
(selectedLunWarning in SanStorageModelBase) but was mistakenly not added
to FcpStorageView before.
This warning message is used to warn about adding a lun that is already
a part of a storage domain. It worked with ISCSI luns, now it also
works with FCP luns.

Change-Id: I5d6b0392977cc228a2a71346d70a1498631ecadd
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1221189
Signed-off-by: Idan Shaby <ishaby@redhat.com>