core: fetch vm dynamic info from vm analyzers
Change the way VmsMonitoring fetches VM dynamic info  from VM
analyzers: iterate the analyzers using the stream API and get
the dynamic info instead of the statistics being set inside
VmsMonitoring by the analyzers.

Note that the added code for processVmsWithDevicesChange is
not as efficient as it used to be, but the process of extracting
the handling of the devices from VmsMonitoring is in progress
and is expected to change this code anyway.

Change-Id: I4aa8e5a2778c0aae85644043b3a1530970a862eb
Signed-off-by: Arik Hadas <ahadas@redhat.com>