core: retrieve vm-jobs only if needed
Previously, we queried vm-jobs from the database every time that
VMs-monitoring was executed - even when no VM reported vm-jobs.
Now, the vm-jobs are queried only if a VM reports VM-jobs.

The query is now done by VmAnalyzer instead of VmsMonitoring.

The drawback of this approach is that when we poll host with 100
running VMs, we will query the database 100 times in worse case
(when all VMs report VM-jobs) instead of only once. However, the
common case is that no VM or only a few report VM-jobs and thus
this approach is better for the common case. For the extreme case
a mechanism that caches the jobs instead of querying the database
everytime would be the best solution.

Change-Id: I2b9ee3c1bb0c6d7316a1a4f73488a550a1f994e5
Signed-off-by: Arik Hadas <ahadas@redhat.com>