core: Reimplement PKIResources.Resource as class
PKIResources.Resource was implemented as enum, but we used Config value
as a parameter for enum constructor. This caused troubles because enum
constructor was called during class loading but Config values are
initialized during Backend EJB initialization, so if some called
PKIResourceServlet to get certificate before Backend was initialized,
enum was initialized incorrectly and servlet returned error until JBoss
instance was restarted.

  1. Connection refused - before JBoss instance starts up
  2. 404 Not Found - JBoss instance is up, but engine application is
     not yet deployed
  3. 500 Internal Server Error - in server log there's our custom
     message describing that certificate is not available yet as engine
     startup hasn't been finished yet
  4. Success, certificate returned - JBoss instance is up, engine
     application fisnished initialization

Change-Id: If5fea4db033bddb5103cd412627e2506d5167a3d
Bug-Url: https://bugzilla.redhat.com/1294025
Signed-off-by: Martin Perina <mperina@redhat.com>